feat(OUT-1221): add script to backfill task assigned activity logs
